Runnable jar file doesn't work
Hello Together
I still have a small problem I wanted to create an executable program for my app, so that people without Eclipse can start the app, too. So either export as exe or Jar file. After some googling I found out that in eclipse a program can be exported as Jar file. This is what I did. But when I try to start the jar file I get an error message which is probably related to my user library JavaFX. It seems that this error message can not be given when exporting. How could you solve this when exporting?

Can anyone generate from my Applikation a runnable jar file? That would be very nice.

Thanks a lot,
Rubberduck